// apps/web-v4/src/app/models/certification.model.ts
export interface ICertificationResponse {
  resultList: ICertification;
}
export interface ICertification {
  passedList: IContent[];
  canAttemptList: IContent[];
  cannotAttemptList: IContent[];
  ongoingList: IContent[];
  sortedList: IContent[];
}
export interface ICertificationRequest {
  userEmail: string;
  tracks: string[];
  sortOrder?: 'asc' | 'desc';
}